src-dashboard/requirements.txt
dj-database-url==0.5.0
Django==1.11.25
django-chartjs==1.5.0
django-classy-tags==0.9.0
django-cms==3.6.0
django-filer==1.5.0
django-formtools==2.1
django-js-asset==1.2.2
django-mptt==0.10.0
django-polymorphic==2.0.3
django-sekizai==1.0.0
django-static-jquery==2.1.4
django-treebeard==4.3
djangocms-admin-style==1.4.0
djangocms-attributes-field==1.1.0
djangocms-charts==1.3.0
djangocms-column==1.9.0
djangocms-file==2.3.0
djangocms-googlemap==1.3.0
djangocms-installer==1.1.1
djangocms-link==2.5.0
djangocms-picture==2.1.3
djangocms-snippet==2.1.0
djangocms-style==2.2.0
djangocms-text-ckeditor==3.7.0
djangocms-video==2.1.1
gunicorn==19.9.0
html5lib==1.0.1
Pillow==6.2.0
psycopg2-binary==2.8.3
pytz==2019.3
six==1.12.0
tzlocal==2.0.0
Unidecode==1.0.23
webencodings==0.5.1
